fix: moved db config from package to repo level
